requirements.txt
vk-api
gigachat
requests
python-dotenv
openai-whisper
ffmpeg-python
librosa
pydub
numpy
